Expand SimControl commands, and use better random numbers

+++ SimPort/SimPortConf.h
@@ -59,17 +59,17 @@ public:
 	{}
 
 	std::vector<uint32_t> BinaryIndicies;
-	std::map<size_t, bool> BinaryStartVals;
-	std::map<size_t, bool> BinaryForcedStates;
-	std::map<size_t, unsigned int> BinaryUpdateIntervalms;
+	std::map<uint32_t, bool> BinaryStartVals;
+	std::map<uint32_t, bool> BinaryForcedStates;
+	std::map<uint32_t, unsigned int> BinaryUpdateIntervalms;
 	std::vector<uint32_t> AnalogIndicies;
-	std::map<size_t, double> AnalogStartVals;
-	std::map<size_t, bool> AnalogForcedStates;
-	std::map<size_t, unsigned int> AnalogUpdateIntervalms;
-	std::map<size_t, double> AnalogStdDevs;
+	std::map<uint32_t, double> AnalogStartVals;
+	std::map<uint32_t, bool> AnalogForcedStates;
+	std::map<uint32_t, unsigned int> AnalogUpdateIntervalms;
+	std::map<uint32_t, double> AnalogStdDevs;
 	std::vector<uint32_t> ControlIndicies;
-	std::map<size_t, unsigned int> ControlIntervalms;
-	std::map<size_t, std::vector<BinaryFeedback>> ControlFeedback;
+	std::map<uint32_t, unsigned int> ControlIntervalms;
+	std::map<uint32_t, std::vector<BinaryFeedback>> ControlFeedback;
 
 	double default_std_dev_factor;
 };
